# lesson3/src/utils/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "title": "SEPT-2022",
    "description": "NODEJS MODULE FOR SEPT-2022",
    "contact": {
      "email": "lidiyakocherzchuk@gmail.com"
    }
  },
  "paths": {
...
    "/auth/password/change": {
      "patch": {
        "tags": [
          "Auth"
        ],
        "summary": "Change password",
        "operationId": "changePassword",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "schema": {
              "type": "string",
              "default": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfdXNlcl9pZCI6IjY0NGFjZmFiMmE4NjlhOWNlNzMyYThkMiIsInVzZXJOYW1lIjoiTGlkYSIsImlhdCI6MTY4MjYyNDcyNiwiZXhwIjoxNjgyODgzOTI2fQ.NqFIBWjoQR91kinHa9tpmdjM3HHxFwj7AK0rSwMJRL4"
            }
          },
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "default": {
                "email": "lida@gmail.com",
                "password": "1235LoVe*"
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Changed password successfully"
          },
          "401": {
            "description": "No token provided"
          },
          "400": {
            "description": "Bed request"
          }
        }
      }
    },
    "/auth/password/forgot": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Forgot password",
        "operationId": "forgotPassword",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "default": {
                "email": "lida@gmail.com"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password change request sent successfully"
          },
          "400": {
            "description": "Bed request"
          }
        }
      }
    },
    "/auth/password/forgot/{token}": {
      "patch": {
        "tags": [
          "Auth"
        ],
        "summary": "Set forgot password",
        "operationId": "setForgotPassword",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "boby",
            "required": true,
            "schema": {
              "type": "object",
              "default": {
                "password": "1235LoVe*"
              }
            }
          },
          {
            "in": "path",
            "name": "token",
            "required": true,
            "schema": {
              "type": "string",
              "default": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfdXNlcl9pZCI6IjY0NGFjZmFiMmE4NjlhOWNlNzMyYThkMiIsInVzZXJOYW1lIjoiTGlkYSIsImlhdCI6MTY4MjYyNDcyNiwiZXhwIjoxNjgyODgzOTI2fQ.NqFIBWjoQR91kinHa9tpmdjM3HHxFwj7AK0rSwMJRL4"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Password changed successfully"
          },
          "401": {
            "description": "No token provided"
          },
          "400": {
            "description": "Bed request"
          }
        }
      }
    },
...
    "/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Refresh",
        "operationId": "refresh",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": true,
            "schema": {
              "type": "string",
              "default": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfdXNlcl9pZCI6IjY0NGFjZmFiMmE4NjlhOWNlNzMyYThkMiIsInVzZXJOYW1lIjoiTGlkYSIsImlhdCI6MTY4MjYyNDcyNiwiZXhwIjoxNjgyODgzOTI2fQ.NqFIBWjoQR91kinHa9tpmdjM3HHxFwj7AK0rSwMJRL4"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Token"
            }
          },
          "401": {
            "description": "No token provided"
          },
          "400": {
            "description": "Bed request"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Token": {
      "type": "object",
      "properties": {
        "accessToken": {
          "type": "string",
          "example": "sdglskrtg5f3srg4drgsr5gs"
        },
        "refreshToken": {
          "type": "string",
          "example": "sdglskrtg5f3srg4drgsr5gs"
        }
      }
    }
  }
}